chore: organized nosql schemas to align with DataDict

+++ backend/schemas/market.py
@@ -1,10 +1,10 @@
-from pydantic import BaseModel
+from pydantic import BaseModel, Field
 from typing import Optional
 
 class MarketSearchQuery(BaseModel):
     location_adm2: Optional[str] = None
     housing_type: Optional[str] = None
-    max_budget: Optional[int] = None
+    max_budget: Optional[float] = None
     semantic_query: Optional[str] = None
 
 class Ratings(BaseModel):
